+++ lib/hrk_nasa_apis.dart
@@ -2,7 +2,6 @@ library;
 
 export 'package:dio/dio.dart';
 
-export 'package:hrk_batteries/hrk_batteries.dart'
-    show HttpResponseStatusCode, JsonMap, FromJsonFunction;
+export 'package:hrk_batteries/hrk_batteries.dart';
 
 export 'src/nasa_apis.dart';