+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/close_approach_body.dart
@@ -17,6 +17,8 @@ enum CloseApproachBody {
   uranus,
   @JsonValue('Neptn')
   neptune,
+
+  /// Mentioned in documentation, but returns 400 if used
   @JsonValue('Pluto')
   pluto,
   @JsonValue('Moon')