+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/query_parameters.dart
@@ -1,2 +1,3 @@
 export 'close_approach_body.dart';
 export 'sbdb_cad_query_parameters.dart';
+export 'small_body.dart';