+++ lib/src/services/ssd_cneos/sbdb_cad_api/data/query_parameters/small_body.dart
@@ -1,0 +1,11 @@
+enum SmallBody {
+  pha(displayName: 'PHA'),
+  nea(displayName: 'NEA'),
+  comet(displayName: 'Comet'),
+  neaComet(displayName: 'NEA and Comet'),
+  neo(displayName: 'NEO');
+
+  const SmallBody({required this.displayName});
+
+  final String displayName;
+}